Moved Point, Distance, Line and Function into one file
